Split Hermes cron run history
